+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Release Notes
 
 
+## v1.0.2 (2023-11-28)
+
+- package errors
+- run tanpa build => go run errors.go
+
 ## v1.0.1 (2023-11-28)
 
 - package fmt